transaction-list: Extract TransactionPlayersList module
