+++ frontend/src/components/TransactionList.js
@@ -6,20 +6,7 @@ import {
   fetchAllInvalidTransactions
 } from "../queries";
 import TransactionsSkeleton from "../skeletons/Transactions";
-
-const PlayersList = props => {
-  return (
-    <ul class="w-48 text-sm font-medium text-gray-900 bg-white border border-gray-200 rounded-lg dark:bg-gray-700 dark:border-gray-600 dark:text-white">
-      <For each={props.players}>
-        {player_name => (
-          <li class="w-full px-4 py-2 border-b border-gray-200 rounded-t-lg dark:border-gray-600">
-            {player_name}
-          </li>
-        )}
-      </For>
-    </ul>
-  );
-};
+import TransactionPlayersList from "./TransactionPlayersList";
 
 const TransactionList = props => {
   const [query, setQuery] = createSignal();
@@ -125,7 +112,9 @@ const TransactionList = props => {
                           when={nPlayers !== 1}
                           fallback={transaction.players[0]}
                         >
-                          <PlayersList players={transaction.players} />
+                          <TransactionPlayersList
+                            players={transaction.players}
+                          />
                         </Show>
                       </td>
                       <td class="px-6 py-4">
